Create Main
This is the main part of the code and handles almost everything and contains almost all the functions, but it was just getting too big so I decided to put the rest in other files. Also this is one of my first projects so it is not very well made(I made this almost a year ago) but even though it is one of my first ever projects I would say it looks the best out of all of them.